Write the main Javascript file
This file includes three.js code to set up a scene, complete
with a starfield. The starfield is simply just a sphere in which
the camera is inside of.

The planets will be rendered in the middle of the scene, and are also
three.js spheres with topographical and/or atmosphere maps pulled from
PlanetPixelEmporium hi-res photos layered on top.

Camera controls are also included, and the animation loop includes code
to account for camera movement whenever the user moves his/her mouse.