Create planet objects
Created a three.js mesh for each planet. Each planet is simply a
combination of a geometry (a sphere in this case) and a material
to form a three.js 'mesh'. The materials for each planet consisted
of NASA mercator projections of each planet pulled from PlanetPixelEmporium.com

The Earth has an additional mesh of clouds layered on top of the base mesh,
which rotates at its own speed. The ring geometry for Saturn and Neptune
have yet to be done, but will consist of three.js RingGeometry objects.